+++ package.json
@@ -9,7 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@types/next-auth": "^3.15.0",
     "next": "15.0.3",
     "next-auth": "^4.24.10",
     "primeflex": "^3.3.1",
