+++ src/types/next-auth.d.ts
@@ -1,4 +1,5 @@
-// import NextAuth from "next-auth";
+// eslint-disable-next-line
+import NextAuth from "next-auth";
 
 declare module "next-auth" {
   interface Session {
